Migrate Hamcrest asserts to Truth #cleanup
	Change on 2018/08/20 by idv <idv@google.com>

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=209444712

+++ util/src/test/java/com/google/cloud/hadoop/util/ResilientOperationTest.java
@@ -84,7 +84,7 @@ public class ResilientOperationTest {
     assertThat(callTester.timesCalled()).isEqualTo(3);
     verifySleeper(sleeper, 2);
   }
-  
+
   @Test
   public void testCallRetriesAndFailsWithSocketErrors() throws Exception {
     MockSleeper sleeper = new MockSleeper();
@@ -152,8 +152,7 @@ public class ResilientOperationTest {
     BackOff backoff = new RetryBoundedBackOff(3, new BackOffTester());
     assertThat(
             ResilientOperation.retry(
-                    callTester, backoff, RetryDeterminer.DEFAULT, Exception.class, sleeper)
-                .intValue())
+                callTester, backoff, RetryDeterminer.DEFAULT, Exception.class, sleeper))
         .isEqualTo(3);
     assertThat(callTester.timesCalled()).isEqualTo(4);
     verifySleeper(sleeper, 3);
